requirements.txt
accelerate==1.11.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
alembic==1.16.1
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asttokens==3.0.0
attrs==25.4.0
beautifulsoup4==4.14.2
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.4.26
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.10.1
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
databricks-sdk==0.57.0
datapizza-ai==0.0.2
datapizza-ai-clients-google==0.0.4
datapizza-ai-clients-openai==0.0.6
datapizza-ai-clients-openai-like==0.0.6
datapizza-ai-core==0.0.6
datapizza-ai-embedders-fastembedder==0.0.4
datapizza-ai-embedders-google==0.0.3
datapizza-ai-embedders-openai==0.0.5
datapizza-ai-parsers-docling==0.0.5
datapizza-ai-vectorstores-qdrant==0.0.4
debugpy==1.8.14
decorator==5.2.1
dill==0.4.0
distlib==0.4.0
distro==1.9.0
easyocr==1.7.2
et_xmlfile==2.0.0
executing==2.2.0
Faker==37.11.0
fastapi==0.115.12
filelock==3.18.0
filetype==1.2.0
Flask==3.1.1
fonttools==4.58.2
frozenlist==1.8.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.26.0
google-api-python-client==2.184.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.43.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
grpcio==1.75.1
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hf-xet==1.1.10
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.35.3
hyperframe==6.1.0
identify==2.6.15
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
iniconfig==2.1.0
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.1
jsonlines==4.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kagglehub==0.3.12
kiwisolver==1.4.8
latex2mathml==3.78.1
lazy_loader==0.4
Mako==1.3.10
Markdown==3.8
markdown-it-py==4.0.0
marko==2.2.1
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpire==2.10.2
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
narwhals==2.9.0
nest-asyncio==1.6.0
networkx==3.5
ninja==1.13.0
nodeenv==1.9.1
numpy==2.2.6
omegaconf==2.3.0
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
packaging==24.2
pandas==2.2.3
pandas_flavor==0.7.0
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.3.0
pingouin==0.5.5
platformdirs==4.3.8
plotly==6.3.1
pluggy==1.6.0
polyfactory==2.22.3
portalocker==3.2.0
pre_commit==4.2.0
primp==0.15.0
prompt_toolkit==3.0.51
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pydantic==2.11.5
pydantic-extra-types==2.10.6
pydantic-settings==2.11.0
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
pylatexenc==2.10
pyparsing==3.2.3
pypdfium2==4.30.0
pytest==8.4.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
qdrant-client==1.15.1
rapidocr==3.4.2
readmeai==0.6.3
referencing==0.37.0
regex==2025.9.18
requests==2.32.4
rich==14.2.0
rpds-py==0.27.1
rsa==4.9.1
rtree==1.4.1
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
semchunk==2.2.2
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sklearn-compat==0.1.3
sklearn-pandas==2.2.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.8
SQLAlchemy==2.0.41
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.2
statsmodels==0.14.4
streamlit==1.50.0
structlog==24.4.0
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.6.0
tifffile==2025.10.16
tiktoken==0.4.0
tokenizers==0.22.1
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ucimlrepo==0.0.7
uritemplate==4.2.0
urllib3==2.5.0
uv==0.8.23
uvicorn==0.34.3
virtualenv==20.35.3
watchdog==6.0.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
xarray==2025.6.1
xlsxwriter==3.2.9
yarl==1.22.0
yfinance==0.2.66
zipp==3.23.0
